feat: use message content
