+++ packages/ui/src/features/agent/chat/index.ts
@@ -3,3 +3,11 @@ export { ModernAgentConversation } from "./ModernAgentConversation";
 export * from "./WaitingMessages";
 export * from "./AnimatedThinkingDots";
 export { AgentChart, type AgentChartSpec } from "./AgentChart";
+
+// Message rendering components - for custom message styling
+export { default as AllMessagesMixed, type RenderMessageConfig } from "./ModernAgentOutput/AllMessagesMixed";
+export { default as MessageItem } from "./ModernAgentOutput/MessageItem";
+export { MessageContent, type MessageContentProps, type ArtifactLink } from "./ModernAgentOutput/MessageContent";
+
+// Utilities for message handling
+export { insertMessageInTimeline, isInProgress, getWorkstreamId, DONE_STATES } from "./ModernAgentOutput/utils";
